% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{add}
\alias{add}
\alias{multiply}
\title{Basic calculations}
\usage{
add(x, y)

multiply(x, y)
}
\arguments{
\item{x}{Numeric.}

\item{y}{Numeric.}
}
\value{
The sum or the product of \code{x} and \code{y}.
}
\description{
Some useful operations like + and *
}
\examples{
add(4, 7)
}
