% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irishead}
\alias{irishead}
\title{Iris's head}
\format{
Data frame:
\describe{
\item{(Sepal/Petal).(Length/Width)}{Numeric}
\item{Species}{Factor}
}
}
\usage{
irishead
}
\description{
Iris's head
}
\keyword{datasets}
