% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCpackage-package.R
\docType{package}
\name{LCpackage-package}
\alias{LCpackage}
\alias{LCpackage-package}
\title{LCpackage: This is a Demo for State of the R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

To learn how to create a package with SOTR.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StateOfTheR/LCpackage}
  \item Report bugs at \url{https://github.com/StateOfTheR/LCpackage/issues}
}

}
\author{
\strong{Maintainer}: Antoine Bichat \email{a.bichat@yahoo.fr} (\href{https://orcid.org/0000-0001-6599-7081}{ORCID})

Other contributors:
\itemize{
  \item State of the R [copyright holder]
}

}
\keyword{internal}
